+++ src/eslint-config-adeira/runner/index.js
@@ -1,5 +1,0 @@
-// @flow
-
-const runner = require('@adeira/eslint-runner');
-
-module.exports = runner;